Updated the UI for passcodes, added in the request for current pass before updating to a new one.
Updated receipts to be editable. Running into issues where the body text is needing to be a TextEditor but this gets compacted in size when placed in a scrollview, as well as folders having type-able inputs not being the best method as the apps folders are assigned based on predetermined folders (if a user can input any name it breaks the system). Need to redo the edit screen.